Split auth Store into seperate files
To stop this file becoming too big
